+++ packages/vaadin-grid-pro/src/vaadin-grid-pro-inline-editing-mixin.js
@@ -4,8 +4,8 @@
  * This program is available under Commercial Vaadin Developer License 4.0 (CVDLv4).
  * See <a href="https://vaadin.com/license/cvdl-4.0">the website</a> for the complete license.
  */
-import { animationFrame } from '@polymer/polymer/lib/utils/async.js';
-import { Debouncer } from '@polymer/polymer/lib/utils/debounce.js';
+import { animationFrame } from '@vaadin/component-base/src/async.js';
+import { Debouncer } from '@vaadin/component-base/src/debounce.js';
 import { PolymerElement } from '@polymer/polymer/polymer-element.js';
 
 /**
