+++ packages/vaadin-list-mixin/vaadin-list-mixin.js
@@ -5,8 +5,8 @@
  */
 import { FlattenedNodesObserver } from '@polymer/polymer/lib/utils/flattened-nodes-observer.js';
 import { DirHelper } from '@vaadin/component-base/src/dir-helper.js';
-import { Debouncer } from '@polymer/polymer/lib/utils/debounce.js';
-import { timeOut } from '@polymer/polymer/lib/utils/async.js';
+import { Debouncer } from '@vaadin/component-base/src/debounce.js';
+import { timeOut } from '@vaadin/component-base/src/async.js';
 
 /**
  * A mixin for `nav` elements, facilitating navigation and selection of childNodes.
