+++ packages/vaadin-virtual-list/src/virtualizer-iron-list-adapter.js
@@ -1,5 +1,5 @@
-import { timeOut, animationFrame } from './async.js';
-import { Debouncer, flush } from './debounce.js';
+import { timeOut, animationFrame } from '@vaadin/component-base/src/async.js';
+import { Debouncer, flush } from '@vaadin/component-base/src/debounce.js';
 import { ironList } from './iron-list.js';
 
 // iron-list can by default handle sizes up to around 100000.
